👽️ Change to `FileStream` `ctor` overload without `FileStreamOptions`
The `FileStreamOptions` overload has been added in .NET6, which makes it harder for us to be compatible with older versions of the framework including NET Standard.

This is purely a compatibility change that should not affect the final behavior as we are still providing all the same values via another ctor overload.